+++ inject.go
@@ -1,0 +1,6 @@
+package imbue
+
+// InjectOption is an option that changes the behavior of a call to InjectX().
+type InjectOption interface {
+	applyInjectOption()
+}