+++ internal/generate/main.go
@@ -21,6 +21,8 @@ func main() {
 
 func getGenerator(filename string) func(*jen.File) {
 	switch filepath.Base(filename) {
+	case "inject.gen.go":
+		return generator.GenerateInject
 	case "invoke.gen.go":
 		return generator.GenerateInvoke
 	case "with.gen.go":